chore: rename folder to correct name Header
