+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import Header from './components/TemporaryName/Header';
+import Header from './components/Header/Header';
 import Modules from './components/Modules/Modules';
 import Reasons from './components/Reasons/Reasons';
